requirements.txt
fastapi
uvicorn 
python-dotenv
groq